CHANGE: Refactor Table structure to not rely on plant names
Plant names might change depending on the season. For example we might want to
exchange Plants in the rack but keep the sensors at the same position. For that
I have to decouple sensor position & plant names.

In order to do so, I keep the Iot+Sensor -> Plant Name mapping in the `Plant`
table. When IoT is uploading a new observation, the backend is looking up the
currently configured plant name for a given Iot/Sensor key and adds
this (together with a timestamp) into `Observation`.